Fix issue with PBKDF2_ITERATIONS mock
All unit tests are passing now. I fixed the mocking by making it a real
constant and mocking the import in the test, but @craig pointed out
fixing `beforeAll` to `beforeEach` and leaving it a function would've worked too. Since it was only a function _for_ tests, I like changing the tests so it works with it as a real constant more.