Upgrade all dependencies (#122)
* Try to upgrade all dependencies, nightwatch fail

* Most integration tests passing

Remaining tests manually verified.

* Bump node in workflows

* Few more cleanups

* More tests closer

* Still migrating from jasmine spies

* All unit test migration done

Single test failing with a logical difference, trying to figure out how
it's related to the migration.

* Fix breaking change in `protobufjs`

* Fix issue with PBKDF2_ITERATIONS mock

All unit tests are passing now. I fixed the mocking by making it a real
constant and mocking the import in the test, but @craig pointed out
fixing `beforeAll` to `beforeEach` and leaving it a function would've worked too. Since it was only a function _for_ tests, I like changing the tests so it works with it as a real constant more.

* All integration tests _should_ work now

But I can really test because this is sloppy and leaves resources all
over that eventually make me unable to create new threads.

* All integration tests passing

* Remove innaccurate statement about coverage